player camera
got a camera operator that the player will controll to walk around the world. they use the interface to interact with the ball. the ball will be moved based on the transform of the camera player controls are all done for the most part
need to figre out the networking stuff